+++ client/src/app/routes/todos/todo-dialog/todo-dialog.component.ts
@@ -3,18 +3,20 @@ import { CommonModule } from '@angular/common';
 import { ToDoTask } from '../../../shared/models/todo';
 import { TodosService } from '../../../services/todos.service';
 import { ToastrService } from 'ngx-toastr';
+import { SpinnerComponent } from '../../../core/components/spinner/spinner.component';
 
 @Component({
   selector: 'app-todo-dialog',
   standalone: true,
-  imports: [CommonModule],
   templateUrl: './todo-dialog.component.html',
   styleUrl: './todo-dialog.component.scss',
   host: { '(window:keyup.esc)': 'onClose()' },
+  imports: [CommonModule, SpinnerComponent],
 })
 export class TodoDialogComponent {
   @Input() todoTask: ToDoTask | undefined = undefined;
   @Output() taskDeleted = new EventEmitter<boolean>();
+  isLoading = false;
 
   constructor(
     private todosService: TodosService,
@@ -26,10 +28,11 @@ export class TodoDialogComponent {
     this.taskDeleted.emit(true);
   }
   onDelete() {
+    this.isLoading = true;
     this.todosService.deleteTask(this.todoTask!).subscribe({
       next: (res) => {
-        this.taskDeleted.emit(false);
         this.toastr.success(res.message);
+        this.taskDeleted.emit(false);
       },
     });
   }